Update InvestmentManager.sol
Rolled back the TFP-04 fix (check at line 575) (restored the logic to the state when a referrer could only be set during the first deposit). This change was breaking the logic of the referral system.
• Added a check for the presence of a personal deposit in the claim function. If the user has no personal deposit, the claim transaction will revert with the error NoDeposit. This solution addresses the issue of setting a referrer on the first deposit via reinvestment during claim.

The essence of TFP-04 is that a user cannot set a referrer when making their first deposit through claim. Therefore, rolling back the previous incorrect fix will not cause any issues (this part of the logic has been tested). And disallowing claim without a personal deposit resolves TFP-04, since now, in order to receive rewards and reinvest, the user must make a deposit first.